requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
instagrapi==1.16.14
Pillow==9.0.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycryptodomex==3.12.0
pydantic==1.9.0
PySocks==1.7.1
requests==2.27.1
schedule==1.1.0
typing_extensions==4.0.1
urllib3==1.26.8
